Agrego el trapecio. Pero estoy teniendo unos problemas con el mismo. No estoy pudiendo definir bien los puntos. Alto me esta fallando y no logro ver bien que. Deje una posicion y tamaño por defecto (hardcodeado) por ahora para que vean como queda. Del parser HAY QUE ARREGLAR: getInclinacionObjeto() que no esta funcionando bien.
HAY QUE AGREGAR:
getBaseMayorObjeto() y getBaseMenorObjeto() para el trapecio.

+++ tallerDeProgramacionI-SnowBros/headers/Modelo/Objetos/Figuras/Paralelogramo.h
@@ -1,13 +1,12 @@
 #ifndef PARALELOGRAMO_H_
 #define PARALELOGRAMO_H_
 
-#include <Box2D/Box2D.h>
 #include "Figura.h"
 
 class Paralelogramo : public Figura{
 private:
-	int ancho;
-	int alto;
+	float ancho;
+	float alto;
 	int inclinacion;
 public:
 	Paralelogramo(JsonParser *parser, int index, b2World* world);
